# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  name: string;
  nameBn: string;
  icon: string;
  count: number;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  originalPrice: number;
  discount: number;
  rating: number;
  reviews: number;
  image: string;
  seller: string;
  badge: string;
  stock: number;
  description?: string;
}

export interface Review {
  id: number;
  author: string;
  rating: number;
  text: string;
  date: string;
}

export interface OrderItem {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cod: 'cod',
  bkash: 'bkash',
  nagad: 'nagad',
} as const;

export interface OrderInput {
  /** @minItems 1 */
  items: OrderItem[];
  /** @minLength 1 */
  customerName: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 1 */
  address: string;
  paymentMethod: OrderInputPaymentMethod;
}

export interface Order {
  id: string;
  status: string;
  items: OrderItem[];
  total: number;
  createdAt: string;
  customerName: string;
  phone: string;
  address: string;
  paymentMethod: string;
}

export type ListProductsParams = {
q?: string;
category?: string;
sort?: ListProductsSort;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  popular: 'popular',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
} as const;