# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: catalog
    description: Product discovery operations
  - name: commerce
    description: Cart and order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      summary: List product categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: Search and filter products
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [popular, price_asc, price_desc, rating]
        - name: limit
          in: query
          schema:
            type: number
            default: 12
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [catalog]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /deals:
    get:
      operationId: listDeals
      tags: [catalog]
      summary: List featured deals
      responses:
        "200":
          description: Deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /orders:
    post:
      operationId: createOrder
      tags: [commerce]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [commerce]
      summary: List recent orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Category:
      type: object
      required: [id, name, nameBn, icon, count]
      properties:
        id: { type: string }
        name: { type: string }
        nameBn: { type: string }
        icon: { type: string }
        count: { type: number }
    Product:
      type: object
      required: [id, name, category, price, originalPrice, discount, rating, reviews, image, seller, badge, stock]
      properties:
        id: { type: number }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        discount: { type: number }
        rating: { type: number }
        reviews: { type: number }
        image: { type: string }
        seller: { type: string }
        badge: { type: string }
        stock: { type: number }
        description: { type: string }
    Review:
      type: object
      required: [id, author, rating, text, date]
      properties:
        id: { type: number }
        author: { type: string }
        rating: { type: number }
        text: { type: string }
        date: { type: string }
    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: number }
        quantity: { type: number, minimum: 1 }
    OrderInput:
      type: object
      required: [items, customerName, phone, address, paymentMethod]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        customerName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 5 }
        address: { type: string, minLength: 1 }
        paymentMethod:
          type: string
          enum: [cod, bkash, nagad]
    Order:
      type: object
      required: [id, status, items, total, createdAt, customerName, phone, address, paymentMethod]
      properties:
        id: { type: string }
        status: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total: { type: number }
        createdAt: { type: string }
        customerName: { type: string }
        phone: { type: string }
        address: { type: string }
        paymentMethod: { type: string }