# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: string;
  name: string;
  nameBn: string;
  icon: string;
  count: number;
}