# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: string;
  status: string;
  items: OrderItem[];
  total: number;
  createdAt: string;
  customerName: string;
  phone: string;
  address: string;
  paymentMethod: string;
}