# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  author: string;
  rating: number;
  text: string;
  date: string;
}